examples/Python/ChatAgent/requirements.txt
agent-framework-azure-ai
azure-appconfiguration
azure-appconfiguration-provider
azure-identity